split ipollable header to clientfd/serverfd/filefd headers fix order of inclusions etc
- move some header inclusions around
- rename `CRLF 2` to CRLF_LEN
- replace "\r\n" literal strings with macro
