+++ code/includes/ipollable/ServerFD.hpp
@@ -1,3 +1,20 @@
 #pragma once
 
-#define SERVER_FD_HEADERFILE 1
+#include "ipollable/IPollable.hpp"
+#include "server/Server.hpp"
+#include <cstring>
+
+class ServerFD : public IPollable {
+	public:
+		Server *_server;
+		int     _fd;
+		int     _index;
+
+		ServerFD(Server *server, int fd, int index);
+		~ServerFD();
+
+		void    pollin();
+		void    pollout();
+		int     getFileDescriptor() const;
+		Server *getServer() const;
+};
