+++ code/includes/server/Server.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 #include "config/Config.hpp"
-// #include "ipollable/ClientFD.hpp"
-// #include "ipollable/FileFD.hpp"
+#include "ipollable/ClientFD.hpp"
+#include "ipollable/FileFD.hpp"
 #include "ipollable/IPollable.hpp"
-// #include "ipollable/ServerFD.hpp"
+#include "ipollable/ServerFD.hpp"
 #include "utils/Utils.hpp"
 
 #include <arpa/inet.h>
